requirements.txt
argcomplete==1.12.3
asgiref==3.3.4
boto3==1.17.69
botocore==1.20.69
certifi==2020.12.5
cffi==1.14.5
cfn-flip==1.2.3
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
Django==3.2.2
django-cors-headers==3.7.0
django-extensions==3.1.3
django-rest-knox==4.1.0
django-s3-storage==0.13.4
django-storages==1.11.1
djangorestframework==3.12.4
durationpy==0.5
future==0.18.2
hjson==3.0.2
idna==2.10
jmespath==0.10.0
kappa==0.6.0
pep517==0.10.0
pip-tools==6.1.0
placebo==0.9.0
psycopg2-binary==2.8.6
pycparser==2.20
python-dateutil==2.8.1
python-slugify==5.0.2
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tqdm==4.60.0
troposphere==2.7.1
urllib3==1.26.4
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.52.0
